feat(core): add client backpressure, event-size, and stream introspection ABI
- Add byte-budgeted backpressure to the client queue: mes_client_set/get_max_queue_bytes and mes_client_queued_bytes complement the existing event-count limit, with EventQueue charging packet capacity plus checkpoint storage.
- Add mes_client_set/get_max_event_size so the reader clamps oversized binlog events, mirroring the engine cap and normalization.
- Add mes_client_checksum_enabled to report CRC32 trailer detection for wiring into a raw engine's checksum handling.
- Add mes_client_is_streaming to distinguish drainable terminal state from mes_client_is_connected, avoiding busy-loops on MES_ERR_DISCONNECTED.
- Add TransactionGtidTracker so mes_client_current_gtid advances only after a committed transaction is delivered, documented as a delivery ack rather than a durable one.
- Add allow_public_key_retrieval config opt-in for plaintext caching_sha2 auth, disabled by default.
- Make mes_client_stop synchronous (joins the reader thread) and no longer async-signal-safe.
- Harden protocol packet/socket I/O, state machine buffering, TABLE_MAP metadata parsing, row decoding, and connection validation edge cases.

+++ core/src/binary_util.h
@@ -162,7 +162,7 @@ std::string DecodeDecimal(const uint8_t* data, size_t available, uint8_t precisi
  * @param metadata Type-specific metadata from TABLE_MAP event
  * @return Size of the field in bytes, or 0 for unsupported types
  */
-uint32_t CalcFieldSize(uint8_t col_type, const uint8_t* data, size_t buf_len, uint16_t metadata);
+uint32_t CalcFieldSize(uint8_t col_type, const uint8_t* data, size_t buf_len, uint32_t metadata);
 
 }  // namespace mes::binary
